tests(external secrets): update to new api
